% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjelpefunksjoner.R
\name{abonnementFASETUT}
\alias{abonnementFASETUT}
\title{Gererere månedsrapport for abonnement}
\usage{
abonnementFASETUT(
  rnwFil,
  brukernavn = "ukjent",
  reshID = 0,
  register = "norscir",
  datoFra = Sys.Date() - 400,
  datoTil = Sys.Date()
)
}
\arguments{
\item{rnwFil}{navn på Rnw-fila som skal kjøres ('mndRapp.Rnw')}

\item{brukernavn}{abonnentens brukernavn ved bestilling}

\item{reshID}{\itemize{
\item reshID som abonnenten var logget inn med ved bestilling
}}

\item{datoFra}{\itemize{
\item startdato for data som hentes til bruk i rapporten
}}

\item{datoTil}{\itemize{
\item sluttdato for data som hentes til bruk i rapporten
}}
}
\description{
Gererere månedsrapport for abonnement
}
