% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NSFigFordeling.R
\name{FigFordeling}
\alias{FigFordeling}
\title{Søylediagram som viser fordeling av valgt variabel}
\usage{
FigFordeling(RegData, libkat, outfile = "", valgtVar,
  datoFra = "2010-01-01", datoTil = "2050-01-01", AIS = "", minald = 0,
  maxald = 120, erMann = "", traume = "", egenavd = 1, reshID,
  sml = 1)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{libkat}{- sti til bibliotekkatalog.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{valgtVar}{- Må velges: ... AAis, FAis, Alder, DagerRehab, DagerTilRehab, OpphTot[HosptlDy],
UtTil[PlaceDis], SkadeArsak[Scietiol],  Pustehjelp[VentAssi].}

\item{datoFra}{<- '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{AIS}{- AISgrad ved innleggelse alle(''), velge en eller flere fra A,B,C,D,E,U}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', standard: ikke valgt}

\item{egenavd}{- 1:eget sykehus, 0:hele landet (standard) Kun for valgtVar=='NevrNivaaInnUt'}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{sml}{- Sammenligne med resten av landet, standard (1) eller ikke(0). Ikke aktiv når hele landet valgt.
Skipper inputktr for denne
Ikke_med avd - 0: data fra hele landet, 1: data fra egen avdeling (standard)
valgtMaal - 'Med' = median. Alt annet gir gjennomsnitt}
}
\description{
Alder - Alder
OpphTot - Lengde på opphold – totalt (HosptlDy) 20d… >200
Lengde på opphold – rehabilitering (Utskrevet – Innl. rehab dato)
Tid fra skadedato til oppstart rehab [trauma/ikke], 5d interv
AIS, A-E+U, innleggelse og kontroll
Utskrives til [horisontal]
Årsak til skade
}

