% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_norscirFLYTTETtil_nordicscir.R
\name{kjorApp}
\alias{kjorApp}
\title{Run the application}
\usage{
kjorApp(browser = FALSE, logAsJson = FALSE)
}
\arguments{
\item{browser}{Run app in browser}

\item{logAsJson}{Log in json format}
}
\value{
Shiny app
}
\description{
Run the application
}
