% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorScir.R
\name{norscir}
\alias{norscir}
\title{Some R resources needed for NorScir at Rapporteket}
\description{
Some R resources needed for NorScir at Rapporteket
}
\section{Sample data}{

The package contains a set of sample data that can be used for development
purposes. Not to be used in a PRODUCTION environment!
}

\section{Logo}{

Logo is accessed via:
\code{system.file("NorSCIR_logo_blue_norsk.pdf", package = "norscir")}
}

\examples{
library(norscir)
data("NorScirSampleData")
names(NorScirSampleData)

}
