% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NorScirSampleData}
\alias{NorScirSampleData}
\title{NorScir sample data}
\format{
\subsection{\code{NorScirSampleData}}{

A data frame with 560 rows and 104 columns:
}
}
\usage{
NorScirSampleData
}
\description{
A subset of data from NorScir
}
\keyword{datasets}
