% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFigAndelStabel.R
\name{NSFigAndelStabel}
\alias{NSFigAndelStabel}
\title{Søylediagram som sammenligner fordeling (vha. stablede søyler) av valgt variabel}
\usage{
NSFigAndelStabel(RegData, outfile = "", valgtVar, datoFra = "2010-01-01",
  datoTil = "3000-01-01", minald = 0, maxald = 120, erMann = "",
  traume = "", enhetsUtvalg = enhetsUtvalg, reshID, hentData = 1)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}

\item{outfile}{- navn på fil figuren skrives ned til.}

\item{valgtVar}{- Må velges: ... NevrNivaaInnUt, NevrNivaaInn, NevrNivaaUt}

\item{datoFra}{<- '2010-01-01'    # min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', standard: ikke valgt}

\item{enhetsUtvalg}{- 1:eget sykehus, 0:hele landet (standard) Kun for valgtVar=='NevrNivaaInnUt'}

\item{reshID}{- avdelingsid for egen avdeling, standard: 0-hele landet.}

\item{hentData}{Settes til 1 (standard) om data skal lastes i funksjonen.
Settes til en annen verdi om data leveres til funksjonen gjennom 'RegData',
eksempelvis ved bruk av eksempeldatasettet eller ved kall fra andre
funksjoner der data allerede er tilgjengelig.}
}
\description{
Søylediagram som sammenligner fordeling (vha. stablede søyler) av valgt variabel
hos fra to ulike utvalg, f.eks. eget/resten, før/etter
(Kan velge å ikke ha med sammenligninga)
Søylediagram som viser andeler av valgt variabel:
NevrNivaaInn' - Nevrologisk nivå ved innskrivelse, sml. eget, andre
NevrNivaaUt' - Nevrologisk nivå ved utskrivelse, sml. eget, andre
NevrNivaaInnUt' - sml. Nevrologisk nivå ved inn- og utskrivelse
Hvilke kombinasjonsmuligheter har vi?
Inn / Ut (Velge: bare eget/hele landet)
Eget / resten av landet (Velge: Inn eller Ut)
}

