% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NSLibUtvalg.R
\name{NSLibUtvalg}
\alias{NSLibUtvalg}
\title{Støttefunksjon for NorScir}
\usage{
NSLibUtvalg(RegData, datoFra = "2010-01-01", datoTil = "3000-05-25",
  minald = 0, maxald = 120, erMann = "", traume = "", AIS = "",
  fargepalett = "BlaaOff")
}
\arguments{
\item{datoFra}{<- '2010-01-01'. Min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '3000-05-25'}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: ''(alt annet enn 0 og 1), dvs. begge}
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og
utvalgsteksten.
}

