% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NSLoadRegData.R
\name{NSLoadRegData}
\alias{NSLoadRegData}
\title{Provide global dataframe for NorScir}
\usage{
NSLoadRegData()
}
\value{
RegData data frame
}
\description{
Provides NorScir data from staging
}

