% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSPreprosesser.R
\name{NSPreprosesser}
\alias{NSPreprosesser}
\title{Preprosesserer data fra Nordic Scir}
\usage{
NSPreprosesser(RegData)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret.}
}
\value{
Data En liste med det filtrerte datasettet
}
\description{
Denne funksjonen navner om variabler og beregner evt. nye.
}

