% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_norscirFLYTTETtil_nordicscir.R
\name{server_norscirFLYTTET}
\alias{server_norscirFLYTTET}
\title{Server-del til norscir-appen}
\usage{
server_norscirFLYTTET(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
Server-delen til norscir-appen
}
\description{
Server-del til norscir-appen
}
