% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_norscir.R
\name{ui_norscir}
\alias{ui_norscir}
\title{Brukergrensesnitt (ui) til nordscir-appen}
\usage{
ui_norscir()
}
\value{
Brukergrensesnittet (ui) til nordscir-appen
}
\description{
Brukergrensesnitt (ui) til nordscir-appen
}
