% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_norscirFLYTTETtil_nordicscir.R
\name{ui_norscirFLYTTET}
\alias{ui_norscirFLYTTET}
\title{Brukergrensesnitt (ui) til nordscir-appen}
\usage{
ui_norscirFLYTTET()
}
\value{
Brukergrensesnittet (ui) til nordscir-appen
}
\description{
Brukergrensesnitt (ui) til nordscir-appen
}
