% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NSFigMeanMed.R
\name{FigMeanMed}
\alias{FigMeanMed}
\title{Funksjon som genererer en figur med gjennomsnitt/median}
\usage{
FigMeanMed(RegData, valgtVar, valgtMaal = "Gjsn", datoFra = "2010-01-01",
  datoTil = "2050-12-31", AIS = "", minald = 0, maxald = 130,
  erMann = "", traume = "", libkat, outfile = "", hentData = 1)
}
\arguments{
\item{RegData}{Ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Må velges: ... Alder, DagerRehab, DagerTilRehab, OpphTot[HosptlDy],}

\item{valgtMaal}{- 'Med' = median. Alt annet gir gjennomsnitt}

\item{datoFra}{<- '2010-01-01'. Min og max dato i utvalget vises alltid i figuren.}

\item{datoTil}{<- '2013-05-25'}

\item{AIS}{- AISgrad ved innleggelse alle(''), velge en eller flere fra A,B,C,D,E,U}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{traume}{- 'ja','nei', standard: ikke valgt}

\item{libkat}{Sti til bibliotekkatalog (utgår)}

\item{outfile}{- navn på fil figuren skrives ned til}

\item{hentData}{Settes til 1 (standard) om data skal lastes i funksjonen.
Settes til en annen verdi om data leveres til funksjonen gjennom 'RegData',
eksempelvis ved bruk av eksempeldatasettet eller ved kall fra andre
funksjoner der data allerede er tilgjengelig.}
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median for hvert sykehus
og kan ta inn ulike numeriske variable. Funksjonen er delvis skrevet for å
kunne brukes til andre grupperingsvariable enn sykehus.
}

