% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NorScirSampleData.R
\name{NorScirSampleData}
\alias{NorScirSampleData}
\title{Sample dataset for NorScir.}
\format{A data frame with xxx rows and yy variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
}}
\description{
A dataset containing a sample data with nonesense personal info
}

