% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorScirSampleData.R
\name{NorScirSampleData}
\alias{NorScirSampleData}
\title{Sample dataset for NorScir.}
\format{
A data frame with 560 rows and 104 variables:
\describe{
\item{InjuryDt}{Date of injury}
\item{AdmitDt}{...}
\item{DischgDt}{}
\item{ANeuExmDt}{}
\item{FNeuExmDt}{}
\item{QolDt}{}
\item{AdmitRehDt}{}
\item{CNeuExmDt}{}
\item{InjuryDateUnknown}{}
\item{PasientId}{}
\item{isMale}{}
\item{AlderAar}{}
\item{ReshId}{}
\item{ShNavn}{}
\item{DataSett}{}
\item{RelatedID}{}
\item{SkjemaID}{}
\item{DatasettGuid}{}
\item{RapportgrunnlagID}{}
\item{KontaktID}{}
\item{DataSettID}{}
\item{Scietiol}{}
\item{isVrtbrInj}{}
\item{isAssocInj}{}
\item{isSpnlSurg}{}
\item{VentAssi}{}
\item{PlaceDis}{}
\item{ASensLvlAreaL}{}
\item{ASensLvlLC}{}
\item{ASensLvlLT}{}
\item{ASensLvlLL}{}
\item{ASensLvlLS}{}
\item{ASensLvlAreaR}{}
\item{ASensLvlRC}{}
\item{ASensLvlRT}{}
\item{ASensLvlRL}{}
\item{ASensLvlRS}{}
\item{AMtrLvlAreaL}{}
\item{AMtrLvlLC}{}
\item{AMtrLvlLT}{}
\item{AMtrLvlLL}{}
\item{AMtrLvlLS}{}
\item{AMtrLvlAreaR}{}
\item{AMtrLvlRC}{}
\item{AMtrLvlRT}{}
\item{AMtrLvlRL}{}
\item{AMtrLvlRS}{}
\item{AAis}{}
\item{FSensLvlAreaL}{}
\item{FSensLvlLC}{}
\item{FSensLvlLT}{}
\item{FSensLvlLL}{}
\item{FSensLvlLS}{}
\item{FSensLvlAreaR}{}
\item{FSensLvlRC}{}
\item{FSensLvlRT}{}
\item{FSensLvlRL}{}
\item{FSensLvlRS}{}
\item{FMtrLvlAreaL}{}
\item{FMtrLvlLC}{}
\item{FMtrLvlLT}{}
\item{FMtrLvlLL}{}
\item{FMtrLvlLS}{}
\item{FMtrLvlAreaR}{}
\item{FMtrLvlRC}{}
\item{FMtrLvlRT}{}
\item{FMtrLvlRL}{}
\item{FMtrLvlRS}{}
\item{FAis}{}
\item{Skjematype}{}
\item{CSensLvlAreaL}{}
\item{CSensLvlLC}{}
\item{CSensLvlLT}{}
\item{CSensLvlLL}{}
\item{CSensLvlLS}{}
\item{CSensLvlAreaR}{}
\item{CSensLvlRC}{}
\item{CSensLvlRT}{}
\item{CSensLvlRL}{}
\item{CSensLvlRS}{}
\item{CMtrLvlAreaL}{}
\item{CMtrLvlLC}{}
\item{CMtrLvlLT}{}
\item{CMtrLvlLL}{}
\item{CMtrLvlLS}{}
\item{CMtrLvlAreaR}{}
\item{CMtrLvlRC}{}
\item{CMtrLvlRT}{}
\item{CMtrLvlRL}{}
\item{CMtrLvlRS}{}
\item{CAis}{}
\item{ANeuNoMeasure}{}
\item{FNeuNoMeasure}{}
\item{CNeuNoMeasure}{}
\item{OutOfHosptlDy}{}
\item{HosptlDy}{}
\item{SatGenrl}{}
\item{SatPhys}{}
\item{SatPsych}{}
\item{DagerRehab}{}
\item{DagerTilRehab}{}
\item{NevrNivaaInn}{}
\item{NevrNivaaUt}{}
\item{NevrNivaaKtr}{}
}
}
\description{
A dataset containing a sample data with nonesense personal info.
For testing purposes. Someone please provide variable descriptions...
}
